Resolved remaining linter errors
