+++ lib/moonchild.js
@@ -5,7 +5,7 @@ var parser = require('./metadata');
 var estraverse = require('estraverse');
 var expanders = require('../third_party/expanders');
 var util = require('./util');
-var Channel = require('../common/channel');
+var Channel = require('../lib/channel');
 
 var globalHooks = {};
 var globalExtensions = {};
